# This is a combination of 4 commits. # The first commit's message is:
# This is a combination of 6 commits.
# The first commit's message is:

Updating api model in acs-generator to not have apiversion on the types used by acs apis. but still accepting them in the json for commandline interface. So that json sent to arm is the same as json given to command line and we won't have unintialized fields in our api code when arm doesn't send it to us.

# This is the commit message #2:

Updating api model in acs-generator to not have apiversion on the types used by acs apis. but still accepting them in the json for commandline interface. So that json sent to arm is the same as json given to command line and we won't have unintialized fields in our api code when arm doesn't send it to us.

Updating api model in acs-generator to not have apiversion on the types used by acs apis. but still accepting them in the json for commandline interface. So that json sent to arm is the same as json given to command line and we won't have unintialized fields in our api code when arm doesn't send it to us.

test

# The commit message #3 will be skipped:

# Updating api model in acs-generator to not have apiversion on the types used by acs apis. but still accepting them in the json for commandline interface. So that json sent to arm is the same as json given to command line and we won't have unintialized fields in our api code when arm doesn't send it to us.
#
# Updating api model in acs-generator to not have apiversion on the types used by acs apis. but still accepting them in the json for commandline interface. So that json sent to arm is the same as json given to command line and we won't have unintialized fields in our api code when arm doesn't send it to us.

# The commit message #4 will be skipped:

# test

# The commit message #5 will be skipped:

# Updating api model in acs-generator to not have apiversion on the types used by acs apis. but still accepting them in the json for commandline interface. So that json sent to arm is the same as json given to command line and we won't have unintialized fields in our api code when arm doesn't send it to us.

# The commit message #6 will be skipped:

# fixing some build errors

# The commit message #2 will be skipped:

# test

# The commit message #3 will be skipped:

# test

# The commit message #4 will be skipped:

# test